transaction start

+++ src/main/java/com/diaspark/INB/DTO/UserAccountDto.java
@@ -4,7 +4,7 @@ public class UserAccountDto {
 	
 
 	private int accountType;
-    private long accountBalance;
+    private double accountBalance;
     private long customerId;
 
 	public long getCustomerId() {
@@ -15,11 +15,11 @@ public class UserAccountDto {
 		this.customerId = customerId;
 	}
 
-	public long getAccountBalance() {
+	public double getAccountBalance() {
 		return accountBalance;
 	}
 
-	public void setAccountBalance(long accountBalance) {
+	public void setAccountBalance(double accountBalance) {
 		this.accountBalance = accountBalance;
 	}
